+++ fakeredis/_basefakesocket.py
@@ -1,7 +1,9 @@
 import itertools
+import operator
 import queue
 import time
 import weakref
+from functools import reduce
 from typing import List, Any, Tuple, Optional, Callable, Union, Match, AnyStr, Generator, Dict
 from xmlrpc.client import ResponseError
 
@@ -33,7 +35,8 @@ def _convert_to_resp2(val: Any) -> Any:
         result = list(itertools.chain(*val.items()))
         return [_convert_to_resp2(item) for item in result]
     if isinstance(val, (list, tuple)):
-        return [_convert_to_resp2(item) for item in val]
+        res = [_convert_to_resp2(item) for item in val]
+        
     return val
 
 
@@ -205,7 +208,7 @@ class BaseFakeSocket:
                 while len(buf) < length + 2:
                     buf += yield
                 fields.append(buf[:length])
-                buf = buf[length + 2 :]  # +2 to skip the CRLF
+                buf = buf[length + 2:]  # +2 to skip the CRLF
             self._process_command(fields)
 
     def _process_command(self, fields: List[bytes]) -> None:
@@ -252,7 +255,7 @@ class BaseFakeSocket:
             self.put_response(result)
 
     def _run_command(
-        self, func: Optional[Callable[[Any], Any]], sig: Signature, args: List[Any], from_script: bool
+            self, func: Optional[Callable[[Any], Any]], sig: Signature, args: List[Any], from_script: bool
     ) -> Any:
         command_items: List[CommandItem] = []
         try:
@@ -390,7 +393,7 @@ class BaseFakeSocket:
                 if match_key(compare_val) and match_type(compare_val):
                     result_data.append(val)
         else:
-            result_data = data[cursor : cursor + count]
+            result_data = data[cursor: cursor + count]
 
         if result_cursor >= len(data):
             result_cursor = 0