+++ docs/about/changelog.md
@@ -13,6 +13,13 @@ toc_depth: 2
 
 - implement support for `WITHSCORE` in `ZRANK` and `ZREVRANK` #357
 - implement support for Connection commands `HELLO`, `CLIENT SETNAME`, `CLIENT GETNAME`, `CLIENT ID`, `CLIENT INFO` #361
+- implement support for `HGETEX`, `HSETEX`, `HGETDEL` #365
+
+### 🐛 Bug Fixes
+
+- Fix `test_tcp_server_started` to close the connection @mgorny (#363)
+- [ACL] Always allow `AUTH` command @wakonig (#362)
+- [ACL] Provide `default` account acl on startup @wakonig (#362)
 
 ## v2.27.0